Safari-609.1.11.1/LayoutTests/fast/dom/Element/getClientRects-return-type-expected.txt

Tests that Element.getClientRects() returns a list of DOMRect objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rects.length is 1
PASS rects[0].__proto__ is DOMRect.prototype
PASS rects[0] is rects.item(0)
PASS successfullyParsed is true

TEST COMPLETE
>
